backend/requirements.txt
asgiref==3.11.1
Django==4.2.30
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.1
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
